# Conflicts:
#	badger_main/Interfaces/PurchaseOrdersRepo.cs
#	badger_view/Controllers/PurchaseOrdersController.cs
#	badger_view/appsettings.json
#	badger_view/wwwroot/js/site.js

+++ badger_main/Models/Sku.cs
@@ -12,7 +12,7 @@ namespace GenericModals.Models
         public string sku { get; set; }
         public int vendor_id { get; set; }
         public int product_id { get; set; }
-        public int weight { get; set; }
+        public decimal weight { get; set; }
         public int created_by { get; set; }
         public int updated_by { get; set; }
         public double created_at { get; set; }
